fix/ front service to show favorite icon
