+++ components/front-service/src/components/AppbarChatIcon.tsx
@@ -1,0 +1,17 @@
+import IconButton from '@mui/material/IconButton';
+import Badge from '@mui/material/Badge';
+import ChatIcon from '@mui/icons-material/Chat';
+
+function AppbarFavoriteIcon() {
+  return (
+    <div>
+      <IconButton size="large" aria-label="show 4 new mails" color="inherit">
+        <Badge badgeContent={4} color="error">
+          <ChatIcon />
+        </Badge>
+      </IconButton>
+    </div>
+  );
+}
+
+export default AppbarFavoriteIcon;
