+++ components/front-service/src/components/AppbarSearchBox.tsx
@@ -1,0 +1,88 @@
+import React, { useState } from 'react';
+import { styled, alpha } from '@mui/material/styles';
+import InputBase from '@mui/material/InputBase';
+import SearchIcon from '@mui/icons-material/Search';
+
+
+const Search = styled('div')(({ theme }) => ({
+    position: 'relative',
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: alpha(theme.palette.common.white, 0.15),
+    '&:hover': {
+      backgroundColor: alpha(theme.palette.common.white, 0.25),
+    },
+    marginRight: theme.spacing(2),
+    marginLeft: 0,
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      marginLeft: theme.spacing(3),
+      width: 'auto',
+    },
+  }));
+  
+const SearchIconWrapper = styled('div')(({ theme }) => ({
+    padding: theme.spacing(0, 2),
+    height: '100%',
+    position: 'absolute',
+    pointerEvents: 'none',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+}));
+
+const StyledInputBase = styled(InputBase)(({ theme }) => ({
+    color: 'inherit',
+    '& .MuiInputBase-input': {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    transition: theme.transitions.create('width'),
+    width: '100%',
+    [theme.breakpoints.up('md')]: {
+        width: '20ch',
+    },
+    },
+}));
+
+function AppbarSearchBox() {
+    const [searchTerm, setSearchTerm] = useState('');
+
+    const handleInputChange = (event) => {
+      setSearchTerm(event.target.value);
+    };
+  
+    const handleKeyPress = (event) => {
+      if (event.key === 'Enter') {
+        callBackendAPI(searchTerm);
+      }
+    };
+
+    const callBackendAPI = async (searchTerm) => {
+        try {
+          const response = await fetch(`/api/search?query=${encodeURIComponent(searchTerm)}`);
+          const data = await response.json();
+          console.log(data); // 
+        } catch (error) {
+          console.error('Error calling backend API:', error);
+        }
+    };
+    
+    return (
+        <div>
+            <Search>
+                <SearchIconWrapper>
+                    <SearchIcon />
+                </SearchIconWrapper>
+                <StyledInputBase
+                    value={searchTerm}
+                    onChange={handleInputChange}
+                    onKeyPress={handleKeyPress}
+                    placeholder="Search User..."
+                    inputProps={{ 'aria-label': 'search' }}
+                />
+            </Search>
+        </div>
+    );
+}
+
+export default AppbarSearchBox;
